Add Builder to PrimitiveSet.
The PrimitiveSet.Builder only builds immutable PrimitiveSet objects.

The interface is a bit different: there is no "setPrimary" method, instead there are two methods for adding primary and non-primary primitives. This simplifies both implementation and usage. It also makes the implementation a bit faster.

The builder becomes unusable after build() has been called. We do this to avoid copying, which makes sure that the builder is as efficient as the current mutable interface.

Because PrimarySet does not enforce that primary is set, we currently also don't enforce it in the builder. This makes migration to the builder easier. Maybe it can be enforced in the future, when keysets are enforced to always have a primary key set.

Also, we remove an unnecessary "protected" keyword.

PiperOrigin-RevId: 455109247
Change-Id: I86ea7eca92f2d5d1a41378d3c41cd57eb82c7744