Modify Internal HPKE KEM Interface to receive interfaces instead of byte arrays for private/public keys for encapsulation and decapsulation. This makes it possible to implement NIST Curves KEMs and better aligns with the HPKE definitions.
Decapsulation is defined in the [RFC](https://www.rfc-editor.org/rfc/rfc9180.html#name-dh-based-kem-dhkem) by the function:
 `def Decap(enc, skR)` skR being the recipient private key. The public key corresponding to the skR is computed as `pk(skR)`

From the RFC:  `The notation pk(skX), depending on its use and the KEM and its implementation, is either the computation of the public key using the private key, or just syntax expressing the retrieval of the public key, assuming it is stored along with the private key object.`

For X25519, we compute the public key once and store along side the private in the implementation, this is a little more efficient than re-calculating it on every decapsulation.

Context on why this is needed for NIST curves:

X25519 perform scalar multiplication on the Montgomery form of the curve, which only requires the `X` coordinate. Each `X` coordinate corresponds to two points, however, the `Y` coordinate isn't required for Montgomery arithmetic and `X` is used as shared secret for the underlying HKDF. Because of this, there's no ambiguity on what to use.

In the NIST curves, similarly, each `X` coordinates also corresponds to two possible Points (`X`, 'Y'), (`X`, `-Y`) (over the field). However, both coordinates are required for the HKDF in HPKE (it uses uncompress form). In the SEC standard, a point can be represented in compressed or uncompressed form, to be able to translate in between avoiding ambiguity, an extra byte is appended to specify the wether it corresponds to the positive or negative (modulo in the field) coordinate.

PiperOrigin-RevId: 461044547
Change-Id: I70ef3bb9dbf5d75cd94a7bedbd17088fc97ecb0b