Reject invalid key id values in JsonKeysetReader in Java.
Non-uint32 key id values are always rejected in C++, Go and Python. In Java we are a bit less strict and also accept negative signed int32 values, because previous versions of the JsonKeysetWriter used to write these and we don't want to break old keysets with such values.

But we don't want to accept any other values, since it is not clearly defined to what value they should be mapped to, and we want to be as consistent as possible between different languages.

It is unlikely that this stricter parsing will break any user, since no implementation of the JsonKeysetWriter in any language has ever created such invalid key ids, and the parsing in all other languages is even stricter than this.

PiperOrigin-RevId: 473985290
Change-Id: I287562b3cab68b6e60a5f2ba487c3848c229bf28