Reject JWTs with duplicate map-keys in JSON objects in Java.
Such tokens are already rejected in C++ and Go (and soon in Python), and we also want to reject them in Java to be consistent between the languages. So this change should be considered a bug fix.

This may potentially break some users that depend on this, but this is unlikely.

This uses the new, Tink-internal JsonParser. That parser uses the Gson JsonReader and is very similar to the Gson JsonParser, but stricter: it rejects duplicated map-keys, checks that strings are valid, and limits the nesting of JSON objects.

PiperOrigin-RevId: 475537871
Change-Id: I66e9e926d89a57de29368cce87fdc8b18e1294e0