Register HmacProtoSerialization and rewrite MacIntegrationTest.
I renamed it to MacTest because integration is usually used for larger tests that test the integration of several systems.

Most of what MacIntegrationTest does is already covered by MacWrapperTest. I replace it with tests for the whole mac package. These tests only use the public API of Tink, and shows how the mac package can be used.

Since the mac package already has implementations of Parameters and Keys for AesCmac and Hmac, I have also added tests for them.

PiperOrigin-RevId: 477124838
Change-Id: I794fb918986a3bb38b4723de7115f89815227a5f