Rewrite NoSecretKeysetHandleTest.
I want to remove all usages of KeysetManager. But also, since this class is deprecated and was replaced by "KeysetHandle.readNoSecret", I think it is better to simply just use the same test cases as used for KeysetHandle.readNoSecret, and show that it does the same.

In one case, it throws a different error.

PiperOrigin-RevId: 482419196
Change-Id: I2e2d11f9128711b64164dd33e3299e42f6fe4cf7