Let parseKeyWithLegacyFallback only fall back to LegacyProtoKey when no parser exists.
When a parser exists and parsing fails, the function should raise a GeneralSecurityException.

KeysetHandle currently contains unparsed keys that can be invalid and parses them when the user tries to access them using getAt(). Before this change, that function would return a LegacyProtoKey that contains the unparsed, invalid key. This behaviour is unexpected and should be changed. The only reasonable behavior is to raise an exception. But raising a checked exception is not possible, because it would break the API. So we have to convert the exception into an unchecked exception. Note that eventually we want this error to be raised earlier when the keyset handle is constructed, in which case we can raise this as a checked exception.

This change may break users that currently have unparsable keys in their keyset and use getAt. But such keysets are unusable anyways, producing a primitive will always throw an exception. And Tink doesn't allow the user to generate such keysets. So I don't think there is a big risk in breaking anybody.

PiperOrigin-RevId: 484271577
Change-Id: I8a8aa275cbf7ca01a5cc4d8332b0aed7133c6d94