Move some helper functions for elliptic curves into internal.
EllipticCurves.java in subtle has a lot of functions that are specific for certain primitives. And some functions that we will need to implement ECDSA Signature keys are not public, such as the test if a key is on a curve.

This change moves some of the elliptic curve functions into an internal helper class. (We keep a reference of them in subtle to not break the subtle API.)

PiperOrigin-RevId: 486635020
Change-Id: I8520da6be01ab17cf4d2bd9343f6d2cc8ac85fce