Fix race condition in AndroidKeysetManager's build().
The build() function is synchronized, but unfortunately that does not help when there are more than one builder object, which is usually the case. This change now adds a global lock to the build function, to make sure that only one thread can be running the build() function.

Note that this fix has already been added to "EncryptedFile" class in android in:
https://android-review.googlesource.com/c/platform/frameworks/support/+/2257884

PiperOrigin-RevId: 496589296
Change-Id: I6c56e26dbe58f191e592ba51ef56fa105525efa9