Only log message when key is not found in verbose mode.
This logging was improved in https://github.com/google/tink/pull/537, but it might still be a bit confusing if the users sees an exception in the log that is the normal work flow.

So I think it's better to turn this off by default, and make it only log in verbose mode.

At the same time, reverse the if statement to make the function a bit simpler.

PiperOrigin-RevId: 496670085
Change-Id: I0254ac0b714321b85da2dc009a7074337d85763b