Add fake AndroidKeystore provider.
This will help us to better test tink's android keystore package. Injecting a mock keystore object does not work well, since there are places where a new keystore instance is created, which overwrites the mock. And the key generation is not and can't easily be injected because a new instance is created for each new key.

In AndroidKeystoreTest, we add the same tests that we do for the fake provider also for the real AndroidKeyStore provider, which shows that for these use-cases, the two behave the same.

Note that we don't implement a fake CipherSpi. I don't know an easy way to do that.
This is fine, we simply use real (but non android keystore) secret keys and real ciphers for these keys, but it means that we will not be able to test all code path in AndroidKeystoreAesGcm.

PiperOrigin-RevId: 500960566
Change-Id: I9ec36152d685765b03628c3a1c62fdd5041f9702