Fix a race condition in AndroidKeystoreKmsClient.
The getOrGenerateNewAeadKey function first checks if a key exists, and generates a new key if it does not. If two threads do this at the same time, there is a possibility for a race condition that both generate a key, and one of the keys will be overwritten.

Without this fix, the test added here may fail with one of these errors:
java.security.UnrecoverableKeyException: Failed to obtain information about key
java.security.InvalidKeyException: Keystore cannot load the key with ID

PiperOrigin-RevId: 502316891
Change-Id: Ia2d4b5ba893aa9d543356becc669caf8ea468647