Don't create a new master key if a keyset exists and improve error message.
If reading the encrypted keyset fails, we try to read it unencrypted. But if that also fails, it means that the keyset most probably was encrypted. So it is better to throw that exception.

I also had to change testBuilder_invalidEncryptedKeysetProto_shouldThrow because it should not fail on the missing master key, but because of the bad keyset.

PiperOrigin-RevId: 504535766
Change-Id: Ifd210dd8d402b014661e17fd592b2460c5842786