Make calling KeysetHandle.Builder::build twice on the same builder object an error.
Calling it twice leads to very unexpected results: the two keysets can be incompatible (because random key IDs are chosen again).

Hence, users should make a new builder. If this would ever be a performance issue (which I do not expect) we need to find a better API.

I do not expect any user to use this, since the current behavior is so unexpected and users will likely notice that it doesn't behave as expected.

PiperOrigin-RevId: 512876843
Change-Id: I8f4a8686551465b07fc2608ac66fbb082cd821dc