Add serialization for RSA SSA PKCS1 keys in Java.
To encode BigIntegers, we use the minimal two's-complement representation in big-endian byte-order. This is the same encoding as is already used when keys are created in Java, but is different from the encoding used in C++, where the minimal encoding is used.

Some tests are still missing in this change. They will be added in the next change.

PiperOrigin-RevId: 525167535
Change-Id: I235d6d32a93e3ee3828646cd47a041c347e72c6f