Document how Tink handles associatedData='' in AES SIV.
AES SIV allows the user to import a list of associated datas, but Tink only uses a single associated data in its interface. When the user inputs an empty associated data byte sequence, Tink will handle this as a list of one empty byte sequence, and not as an empty list.

PiperOrigin-RevId: 527491521
Change-Id: I4f130a25240e30f4cc7cede7a8b952eef31aa0c9