Add another parameters object corresponding to KeyTemplates.get(s) to the map which compares KeyTemplates.get(s) to corresponding Parameters object.
The one added here (AES_128GCM_RAW) does not have an element in
"PredefinedParameters" (for whatever reason). I suggest that in this case, we
compare here with an element which we create directly from AesGcmParameters.

I will add way more with this pattern later if this looks good to you.

PiperOrigin-RevId: 527849982
Change-Id: I36e6bf2ac7f3352a89759751549a8c6903cba8eb