When constructing a KeysetHandle.Builder from another KeysetHandle, check if the entries exist explicitly.
Currently "KeysetHandle" objects can be invalid (unfortunately). While in the long run we want to get rid of this, one easy way to check validity is to call "KeysetHandle.newBuilder(original).build();".

However, this throws a unchecked exception right now if a Key is parsed in an invalid way. We change things so that this throws a GeneralSecurityException.

PiperOrigin-RevId: 538151313
Change-Id: If6568a70cff785dfc2e014a31eb24bd58f86eefe