Mark com.google.http-client as optional.
This class is only needed for the deprecated "KeysDownloader".

Marking this as optional means that users who depend on it need to declare it themselves. This is not quite backwards compatible, but I think it's by far the least bad option: the class is anyhow deprecated, and I don't expect anyone to look into a crypto library for this class. The alternative is that we force everyone to depend on http-client which seems worse.

[1] https://github.com/tink-crypto/tink-java/blob/e8febe585ceda4d16753cfecba8e04bc7cdbe59c/src/main/java/com/google/crypto/tink/util/KeysDownloader.java#L60

PiperOrigin-RevId: 541652601
Change-Id: Ib2d7ccb689eb77f1ebc68048fc25477123039fe1