Restrict KMS envelope AEAD to only use Tink AEAD key types as DEK.
KMS Envelope AEAD is a simple way to encrypt data. So using custom
implementations of AEAD is not needed for this.

This change will help us simplify our implementation of KMS Envelope AEAD,
because we will be able to remove the dependency on Tink's Registry.

Also, this prevents the user from using the envelope encryption key templates or non-AEAD key templates as DEK template by accident.

PiperOrigin-RevId: 542171515
Change-Id: Ie34182b66c25d4d8431513e4bbe6d2221d4c19fb