Delete KeyDerivationkeyTemplates.
This isn't needed anymore -- we now fully support Parameters. Also, this was never in a release, so I'm sure that it's never used. Note that even if a user uses it, they can just replace

    KeyTemplate keyDerivationTemplate =
        KeyDerivationKeyTemplates.createPrfBasedKeyTemplate(a, b);

with

      PrfBasedKeyDerivationParameters.builder()
            .setPrfParameters((PrfParameters) a.toParameters())
            .setDerivedKeyParameters(b.toParameters())
            .build();

(Typically, one doesn't need to convert it into a KeyTemplate -- if one does, one can just call KeyTemplate.fromProto()).

PiperOrigin-RevId: 547406999
Change-Id: If8c8dc08926ace9a05a729a9f1042814972b50d8