In KeyTemplate, either store a proto or a parameters, and convert them on demand.
This means registration of the respective serializer/parser can be done later. Users might save the object, and then later use it in e.g. Keysethandle.generateNew() and only have the serializer register at the later point. Hence we do it like this.

Unfortunately, this means that we may have to throw a TinkBugException (or something else which is unchecked) since the APIs do not throw anything. Within Google nobody called these APIs outside of Tink, so I am relatively sure that this is no problem.

Within Tink, the only remaining usage is in KeysetManager, and I changed things to throw a GeneralSecurityException there.

PiperOrigin-RevId: 549594684
Change-Id: Ifc24fe0b8b85a3a2c2e0a8f75fe8125f3736154f