Remove validation of the public key.
This isn't relevant anymore (and was always a bit of a minor thing). Let me explain.

One of the issues with the KeyManagers is that there can be invalid key protos. For example, suppose that a proto of type AesGcmKey contains a key of length 1. Then, clearly this is invalid.

The principle "Parse, don’t validate" says one should, when one creates an object, check that it is valid, and not even create it.

However, with protos this isn't possible. So the KeyTypeManagers tried to ensure that whenever they actually parse the proto and make it come to existence, it is also validated.

Because of this, the private key ones were designed to call "Validate" (in the public key manager) when they create a public key. This made things a bit more complicated and was probably a mistake from the beginning (instead, I could have probably done things better by just simply calling PublicKeyManager.validate(...) in each private key manager manually).

Anyhow, we don't need this anymore: we now always go key objects anyhow (in case we have a KeyTypeManager), so things will be validated properly. We can hence remove this.

PiperOrigin-RevId: 580834617
Change-Id: Ie88b958d804e4369a622cde2eea52deb3aa0e2eb