Add tests for AesCtrHmacAead which aren't implied by the Parameter & Key objects.
The parameter and key objects sometimes allow more flexibility than Tink in the default API. Hence, we test that the things which the key manager registers is actually reducing the capabilities.

For AesCtrHmacAead, the only thing which falls under this currently is that the AES key size cannot be 192 bits.

PiperOrigin-RevId: 591849331
Change-Id: I2515a2d34531da90c26c3b95341026d3fcc6ed64