Add ChaCha20Poly1305Jce to aead/internal.
This is an implementation of AEAD using the ChaCha20Poly1305 provided by the JCE.

We use the same optimization as for AES-GCM, see discussion in https://github.com/google/tink/issues/208.

The goal is to later add a call to this to the key manager, with a fall-back to the Tink implementation if the JCE doesn't support it.

PiperOrigin-RevId: 600808913
Change-Id: I80fd44e2db0e9bdddfa11d05540a6e91f9b97ef3