Use JCE for ChaCha20Poly1305 AEAD primitive if available.
Tink has its own Java implementation of ChaCha20Poly1305. This is needed because in old JDKs and older versions of Android JCE does not provide ChaCha20Poly1305. But Tink's implementation is about 2 times slower than the implementation provided by the JCE, so it is preferable to use the JCE version if available.

PiperOrigin-RevId: 601063258
Change-Id: I3a42b4c521bf0b33a47d10ae7820beeab7633fb8