Implement XChaCha20Poly1305 using the ChaCha20Poly1305 JCE implementation.
It is very similar to ChaCha20Poly1305Jce. I think it is better to not merge them into one class. The code looked pretty ugly when I tried that, and it might be confusing to the reader, since we keep the x and non-x version in a different class everywhere else.

PiperOrigin-RevId: 602635273
Change-Id: I807802ca74dff2bd71849e3184b9fbc53a341707