Keep GeneralSecurityException if registration goes wrong.
This was recently changed to a TinkBugException, but it is better to keep GeneralSecurityException as it is not always a Bug in Tink if this throws an exception: for example, if a user registers their own key manager for a tink type URL, then this may throw an exception.

registerKeyManagerWithFipsCompatibility should not change the state of the registry if it throws an error.

Also, fix linter error by adding a check that idRequirement is not null.

PiperOrigin-RevId: 611448191
Change-Id: Idaed59460ceca1913a7ccf13011a8b93003257c8