Check that plaintextPosition == plaintextSize in reachedEnd.
The method "SeekableByteChannel position(long newPosition)" only sets "plaintextPosition", and ignores the other internal values.

It's not really clear if we should update those: in particular, one stores the ciphertextSegment, the plaintextSegment (which contain the bytes for segment number currentSegmentNr). We could also invalidate the currentSegmentNr and set it to -1 when we set the position -- I don't see which solution is better.

PiperOrigin-RevId: 641816911
Change-Id: Ife0fc4033ee7b4698253cf04d5224cd01407290f