Remove memory allocation and copy of ciphertext in HPKE decryption.
Instead of creating a ciphertext and then concatenating it with a prefix (which will require copying the whole ciphertext) we create the ciphertext with space for the prefix. The prefix can then be added without copying the ciphertext.

The only remaing code-path where the ciphertext is still copied is when InsecureNonceChaCha20Poly1305Jce.isSupported() is false. But this only happens on a some older Android versions (API version 28 and older). I don't think it's worth optimising for them.

PiperOrigin-RevId: 652561240
Change-Id: I1eb7e6a949e42c59d63b5dbc2eabbe701463d1cc