Let RsaSsaPssSignJce and RsaSsaPssVerifyJce throw an exception if sigHash and mgf1Hash are not equal.
This is already enforced by the non-subtle API, and it is recommended in the RFC. There is no good reason for not choosing them to be equal. It is better to enforce them to be equal. Otherwise, we would need to add additional tests, and do this in all languages.

PiperOrigin-RevId: 653256951
Change-Id: I12e33b0bfae66956ccf1436a1517beece40442ad