Rename X25519Jce to X25519Conscrypt, and make it an object.
Our implementation doesn't work with Bouncycastle, and we don't intend to add support. Since the main goal here is to support Conscrypt, it is better to limit this class to only use Conscrypt. This also makes the testing easier.

We also make it an object so that we can keep the same provider instance.

Additinally, we add an interface for X25519, so that we can provide the pure Java implementation from subtle using the same interface.

PiperOrigin-RevId: 691002085
Change-Id: I7b7692bb789637d006d58adeacb8b88a18a98196