When Conscrypt is available, also use it to create the Java EC keys for ECDSA.
Currently, we use the default provider to create the Java EC keys. If this is not Conscrypt, then in every sign/verify operation, the key needs to be converted into a OpenSSL key, which is inefficient.

To do this, we inline the function that creates the Java EC keys. Then we can pass the security provider as a parameter to the key factory. This also allows us to remove some unnecessary conversion and some checks that are not needed.

We also remove the additional check that creates test signature and a verifier that verifies this test signature. The EcdsaPrivateKey is already validated, so it is not clear what this additional check helps.

This change makes signing small inputs with P256 about twice as fast. The performance of verification improves only by a few percent. Creating a signer get faster, but creating a verifier object gets slower.

PiperOrigin-RevId: 723078454
Change-Id: Idaa3d60606ec20443cd01ade553bab23d525d0dd