Remove PrimitiveSet.hasAnnotations().
PrimitiveSet will be deleted and replaced with a interface similar to KeysetHandle. Hence this cannot stay.

IMO the result is even slightly preferable: the previous API suggested there was a difference between the PrimitiveSet having no annotations and the PrimitiveSet having empty annotations. However, in reality the two were the same.

PiperOrigin-RevId: 728628123
Change-Id: I4ef6c9491e2350d4d41e836f6196e83ba4fea653