Avoid PrimitiveSet.Entry.getOutputPrefix() in the AEAD Wrapper.
This isn't in keyset handle (and will not exist as a concept at some point in the future anymore), hence we avoid it.

For this, we add a getOutputPrefix to LegacyProtoKey (where it actually makes sense; this is something which used to be defined). After this, we can simply use getOutputPrefix either of the AeadKey or the LegacyProtoKey.

PiperOrigin-RevId: 729140903
Change-Id: I6a9002f65b7d4e371f3a0bddf109b9ce8806f58a