Change the wrapping to not take the function mapping Entry -> Primitive.
Currently, KeysetHandle does:
```
config.wrap(
        builder.build(),
        annotations,
        entry -> config.getPrimitive(entry.getKey(), inputPrimitiveClassObject),
        classObject);
```

But this is a bit weird: the configuration knows "getPrimitive", and so it should simply use that in wrap itself instead of having it passed in.

Hence, we remove this argument. But this turns out to be the simplest if we also change the first argument to be a KeysetHandleInterface. Hence, the above simply becomes

```
config.wrap(builder.build().getKeysetHandle(), annotations, classObject);
```

Nevertheless, we also need to remove the primitive from the primitive set.

PiperOrigin-RevId: 739085959
Change-Id: I02b2c1818b9d4dbf1cd6124fe63ed9edf72c1bac