Always register AES-GCM-SIV when register is called.
Currently, it only registers if it finds a Cipher, which is only the case on Android and when Conscrypt is installed as JCE provider. We need to change this, because a config may support AES-GCM-SIV in a different way, without installing a JCE provider.

At the same time, we now check that AES-GCM-SIV is available when the AEAD primitive
is created. In some cases, for example when AeadConfigurationV1 is used, the primitive only failed when a function is called. We consider this as a bug fix, and don't expect that any user is affected by this.

PiperOrigin-RevId: 750607784
Change-Id: Ic415c5ff0f6342e2d3b86ae3188d7ba0113dc93f