Simplify AES-EAX.
The OMAC used in AES-EAX is the same as AES-CMAC, where
OMAC(tag, data) = AES-CMAC(tag || data).

So we can simplify the AES-EAX implementation by re-using the CMAC implementation.

Unfortunately, this simplification results in an additional copy of the input. Avoiding this would require us to make the implementation and the interface of PRF-AES-CMAC more complicated, which we think is not worth the effort, since the impact on the overall performance of AES-EAX is small (it get about %5 slower), and AES-EAX is usually not used for high performance applications.

Also, remove the old and outdated section about our plans for AES-EAX.

PiperOrigin-RevId: 772761236
Change-Id: If38419310c5448fe1cd0a183ef2a9df487038588