Use Conscrypt's AES-CMAC implementation when available.
The current AES-CMAC implementation in Tink uses the AES block cipher and implements CMAC in Java. Conscrypt has to go through the JNI with every call, which makes this implementation very slow when Conscrypt is used.

Conscrypt offers a native AES-CMAC implementation, and it's much faster for large inputs. But unfortunately, for small inputs it is slower than the current implementation, and it may cause locking issues.

We therefore decided to keep using the current implementation for small inputs, (<= 64 bytes), and use Concrypt's native implementation for large inputs.

For large inputs, this improves the performance of AES-CMAC by 10x. It also improves the performance of AES-SIV by about 7x, and AES-EAX by 5x, because they both use AES-CMAC. The performance on small inputs remains the same.

The implementation of ChunkedMac is not affected by this change.

PiperOrigin-RevId: 773302983
Change-Id: I47ff17539a97ac4139b12bbfc06bf3f0921d48c9