Remove thread-local variable from ChaCha20Poly1305 implementations.
I ran some performance tests and it turns out that re-using a cipher is not really faster than creating new cipher objects with the provider as an argument.

So I think we should remove thread-local variables, because this will have some benefits:
- it uses less memory.
- it is simpler.

Not caching the provider would be about 20% to 30% slower, so I think it is worth to cache the provider. The cached provider will also be helpful if we decide to add support for primitives with fixed providers, for example that always use Conscrypt.

PiperOrigin-RevId: 813136921
Change-Id: Ie5a0514ee0eeed4f8d793ce97a0b71caf11986bd