Let JWT keys internally wrap normal Tink signature keys.
Note that we don't change the public API, the new methods are all package-private.

Currently, when a JWT primitive is created, the JWT keys are transformed into a normal Tink signature key. And that key is used to create the JWT primitive. But this key transformation is slow, as it will require the validate the newly created key. With this change, we remove the transformation step, and hence primitive creation is more efficient.

Also, this change allows us to simplify the creating of JWT keys, as we don't have to re-implement key validation.

PiperOrigin-RevId: 830424770
Change-Id: Icb0fbe0de6b2e9588f41dc874a22524a907ae21e