Allow reparsing in ConfigurationV0.
When a user parses a keyset, then registers the serialization, then calls getPrimitive(), this previously failed with ConfigurationV0. We allow this for now.

In the future, we might change this, but we probably will first rework how the serialization registration works.

To simplify implementation, we add a boolean "reparseLegacyKeys" to the primitive registry.

Fixes https://github.com/tink-crypto/tink-java/issues/63

PiperOrigin-RevId: 834314456
Change-Id: I16fcec35dd634570b51d44bdcc5ff24f3fa86b43